bigquery/queries/requirements.txt
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cycler==0.10.0
google-api-core==1.25.1
google-auth==1.25.0
google-cloud-bigquery==2.7.0
google-cloud-core==1.5.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
idna==2.10
importlib-metadata==3.4.0
iniconfig==1.1.1
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.0
packaging==20.9
pandas==1.2.1
Pillow==8.1.0
pluggy==0.13.1
proto-plus==1.13.0
protobuf==3.14.0
py==1.10.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rsa==4.7
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
zipp==3.4.0
